refactored project structure to resembles official golang project template
added dockerfile, script cicd and instructions for docker
