+++ cmd/CoffeeMachineRestApiV1/controllers/front.go
