+++ deployments/CoffeeMachineRestApiV1/.dockerignore
@@ -1,0 +1,3 @@
+*.ps1
+*.bat
+.gitignore
