+++ deployments/CoffeeMachineRestApiV1/BuildDocker.bat
@@ -1,0 +1,1 @@
+PowerShell -NoProfile -ExecutionPolicy Bypass -Command "& .\BuildDocker.ps1"
