+++ deployments/CoffeeMachineRestApiV2/.dockerignore
@@ -1,0 +1,3 @@
+*.ps1
+*.bat
+.gitignore
