+++ deployments/CoffeeMachineRestApiV2/Dockerfile
@@ -1,0 +1,10 @@
+FROM golang:1.15.6-alpine AS build
+
+WORKDIR /go/src/github.com/kOstovic/CoffeeMachine
+COPY . ..
+
+RUN CGO_ENABLED=0 go build -o /bin/coffeemachine cmd/CoffeeMachineRestApiV2/main.go 
+
+FROM scratch
+COPY --from=build /bin/coffeemachine /bin/coffeemachine
+ENTRYPOINT ["/bin/coffeemachine"]
