+++ go.mod
@@ -1,3 +1,3 @@
-module coffeeMachine
+module github.com/kOstovic/CoffeeMachine
 
-go 1.14
+go 1.15
